+++ src/snowcli/cli/snowpark/jobs/commands.py
@@ -7,7 +7,6 @@ from snowcli.cli.common.decorators import global_options_with_connection
 from snowcli.cli.common.flags import DEFAULT_CONTEXT_SETTINGS
 from snowcli.cli.snowpark.common import print_log_lines
 from snowcli.cli.snowpark.jobs.manager import JobManager
-from snowcli.cli.stage.manager import StageManager
 from snowcli.output.decorators import with_output
 
 from snowcli.output.types import SingleQueryResult, CommandResult
@@ -27,30 +26,17 @@ def create(
     spec_path: Path = typer.Option(
         ...,
         "--spec-path",
-        "-s",
         help="Path to the `spec.yaml` file containing the job details.",
         file_okay=True,
         dir_okay=False,
         exists=True,
-    ),
-    stage: str = typer.Option(
-        "SOURCE_STAGE",
-        "--stage",
-        "-l",
-        help="Name of the stage in which to create the job.",
     ),
     **options,
 ) -> CommandResult:
     """
     Creates a job to run in a compute pool.
     """
-    stage_manager = StageManager()
-    stage_manager.create(stage_name=stage)
-    stage_manager.put(local_path=str(spec_path), stage_path=stage, overwrite=True)
-
-    cursor = JobManager().create(
-        compute_pool=compute_pool, spec_path=spec_path, stage=stage
-    )
+    cursor = JobManager().create(compute_pool=compute_pool, spec_path=spec_path)
     return SingleQueryResult(cursor)